music/pipeline/models/requirements.txt
faiss_gpu==1.7.2
numpy==2.1.1
pandas==2.2.3
